pyproject.toml
[project]
name = "printer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "paho-mqtt>=2.1.0",
    "pillow>=11.2.1",
    "pydantic>=2.11.4",
    "python-escpos>=3.1",
    "pyusb>=1.3.1",
    "uvicorn>=0.34.2",
]
